fix: reuse delete action modal in explorer
