+++ packages/frontend/src/components/Explorer/VisualizationCard/VisualizationCard.tsx
@@ -2,19 +2,16 @@ import {
     Button,
     ButtonGroup,
     Card,
-    Classes,
     Collapse,
-    Dialog,
     Divider,
     H5,
     Menu,
     MenuItem,
 } from '@blueprintjs/core';
 import { Popover2 } from '@blueprintjs/popover2';
-import { ChartType, DashboardBasicDetails } from 'common';
+import { ChartType } from 'common';
 import { FC, useState } from 'react';
-import { Link, useHistory, useLocation, useParams } from 'react-router-dom';
-import { getDashboards } from '../../../hooks/dashboard/useDashboards';
+import { useLocation } from 'react-router-dom';
 import {
     useAddVersionMutation,
     useDeleteMutation,
@@ -24,6 +21,7 @@ import { useExplorer } from '../../../providers/ExplorerProvider';
 import BigNumberConfigPanel from '../../BigNumberConfig';
 import ChartConfigPanel from '../../ChartConfigPanel';
 import { ChartDownloadMenu } from '../../ChartDownload';
+import DeleteActionModal from '../../common/modal/DeleteActionModal';
 import LightdashVisualization from '../../LightdashVisualization';
 import VisualizationProvider from '../../LightdashVisualization/VisualizationProvider';
 import AddTilesToDashboardModal from '../../SavedDashboards/AddTilesToDashboardModal';
@@ -31,8 +29,6 @@ import CreateSavedQueryModal from '../../SavedQueries/CreateSavedQueryModal';
 import VisualizationCardOptions from '../VisualizationCardOptions';
 
 const VisualizationCard: FC = () => {
-    const history = useHistory();
-    const { projectUuid } = useParams<{ projectUuid: string }>();
     const [isQueryModalOpen, setIsQueryModalOpen] = useState<boolean>(false);
     const [isAddToDashboardModalOpen, setIsAddToDashboardModalOpen] =
         useState<boolean>(false);
@@ -51,14 +47,11 @@ const VisualizationCard: FC = () => {
         queryResults,
         actions: { setPivotFields, setChartType, setChartConfig },
     } = useExplorer();
-    const { mutate: deleteData, isLoading: isDeleting } = useDeleteMutation();
+    const { isLoading: isDeleting } = useDeleteMutation();
     const update = useAddVersionMutation();
     const [vizIsOpen, setVizisOpen] = useState<boolean>(!!savedChart?.uuid);
     const chartId = savedChart?.uuid || '';
     const { mutate: duplicateChart } = useDuplicateMutation(chartId);
-    const [relatedDashboards, setRelatedDashboards] = useState<
-        DashboardBasicDetails[]
-    >([]);
 
     const searchParams = new URLSearchParams(location.search);
 
@@ -195,24 +188,11 @@ const VisualizationCard: FC = () => {
                                                         icon="trash"
                                                         text="Delete"
                                                         intent="danger"
-                                                        onClick={() => {
-                                                            getDashboards(
-                                                                projectUuid,
-                                                                savedChart?.uuid,
-                                                            ).then(
-                                                                (
-                                                                    dashboards,
-                                                                ) => {
-                                                                    setRelatedDashboards(
-                                                                        dashboards,
-                                                                    );
-
-                                                                    setIsDeleteDialogOpen(
-                                                                        true,
-                                                                    );
-                                                                },
-                                                            );
-                                                        }}
+                                                        onClick={() =>
+                                                            setIsDeleteDialogOpen(
+                                                                true,
+                                                            )
+                                                        }
                                                     />
                                                 </Menu>
                                             }
@@ -253,84 +233,20 @@ const VisualizationCard: FC = () => {
                     onClose={() => setIsAddToDashboardModalOpen(false)}
                 />
             )}
-            <Dialog
-                isOpen={isDeleteDialogOpen}
-                icon="delete"
-                onClose={() =>
-                    !isDeleting ? setIsDeleteDialogOpen(false) : undefined
-                }
-                title={'Delete chart'}
-            >
-                <div className={Classes.DIALOG_BODY}>
-                    <p>
-                        Are you sure you want to delete the chart{' '}
-                        <b>"{chartName}"</b> ?
-                    </p>
-
-                    {relatedDashboards && relatedDashboards.length > 0 && (
-                        <>
-                            <b>
-                                This action will remove a chart tile from{' '}
-                                {relatedDashboards.length} dashboard
-                                {relatedDashboards.length > 1 ? 's' : ''}:
-                            </b>
-                            <ul>
-                                {relatedDashboards.map((dashboard) => {
-                                    return (
-                                        <li>
-                                            <Link
-                                                target="_blank"
-                                                to={`/projects/${projectUuid}/dashboards/${dashboard.uuid}`}
-                                            >
-                                                {dashboard.name}
-                                            </Link>
-                                        </li>
-                                    );
-                                })}
-                            </ul>
-                        </>
-                    )}
-                </div>
-                <div className={Classes.DIALOG_FOOTER}>
-                    <div className={Classes.DIALOG_FOOTER_ACTIONS}>
-                        <Button
-                            disabled={isDeleting}
-                            onClick={() => setIsDeleteDialogOpen(false)}
-                        >
-                            Cancel
-                        </Button>
-                        <Button
-                            disabled={isDeleting}
-                            intent="danger"
-                            onClick={() => {
-                                /*
-                                Check the location of `goBack` after deleting a chart
-                                if we land on an unsaved chart,
-                                we go back further into an empty explore
-                                */
-                                history.listen((loc, action) => {
-                                    if (action === 'POP') {
-                                        if (loc.pathname.includes('/tables/')) {
-                                            history.push(
-                                                `/projects/${projectUuid}/tables`,
-                                            );
-                                        }
-                                    }
-                                });
-
-                                if (savedChart?.uuid) {
-                                    deleteData(savedChart.uuid);
-                                    history.push('/');
-                                }
-
-                                setIsDeleteDialogOpen(false);
-                            }}
-                        >
-                            Delete
-                        </Button>
-                    </div>
-                </div>
-            </Dialog>
+            {isDeleteDialogOpen && (
+                <DeleteActionModal
+                    isOpen={isDeleteDialogOpen}
+                    onClose={() => {
+                        return !isDeleting
+                            ? setIsDeleteDialogOpen(false)
+                            : undefined;
+                    }}
+                    uuid={savedChart?.uuid || ''}
+                    name={chartName || ''}
+                    isChart
+                    isExplorer
+                />
+            )}
         </>
     );
 };
